upload db

+++ pod/ZXKitLoggerWindow.swift
@@ -26,13 +26,41 @@ extension String{
     }
 }
 
+enum PickerType {
+    case share
+    case upload
+}
+
 class ZXKitLoggerWindow: UIWindow {
+    required init?(coder: NSCoder) {
+        super.init(coder: coder)
+    }
+
+    @available(iOS 13.0, *)
+    override init(windowScene: UIWindowScene) {
+        super.init(windowScene: windowScene)
+        self._init()
+    }
+
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+        self._init()
+    }
+
     private var mLogDataArray = [ZXKitLoggerItem]()  //输出的日志信息
     private var mFilterIndexArray = [IndexPath]()   //索引的排序
     private var mCurrentSearchIndex = 0             //当前搜索到的索引
     private var mFileDateNameList = [String]()      //可以分享的文件列表
     private var mShareFileName = ""                 //选中去分享的文件名
-
+    private var pickerType: PickerType = .share {
+        willSet {
+            if newValue == .share {
+                mPickerTipLabel.text = "Please select the log to share".ZXLocaleString
+            } else if newValue == .upload {
+                mPickerTipLabel.text = "Please select the log to upload".ZXLocaleString
+            }
+        }
+    }
     override var isHidden: Bool {
         willSet {
             super.isHidden = newValue
@@ -122,41 +150,6 @@ class ZXKitLoggerWindow: UIWindow {
         }
     }
 
-
-    required init?(coder: NSCoder) {
-        super.init(coder: coder)
-    }
-    
-    @available(iOS 13.0, *)
-    override init(windowScene: UIWindowScene) {
-        super.init(windowScene: windowScene)
-        self._init()
-    }
-    
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        self._init()
-    }
-
-    func insert(model: ZXKitLoggerItem) {
-        if ZXKitLogger.maxDisplayCount != 0 && self.mLogDataArray.count > ZXKitLogger.maxDisplayCount {
-            self.mLogDataArray.removeFirst()
-        }
-        self.mLogDataArray.append(model)
-        if !self.isHidden {
-            self._reloadFilter(model: model)
-        }
-    }
-
-    //
-    func cleanDataArray() {
-        self.mLogDataArray.removeAll()
-        self.mFilterIndexArray.removeAll()
-        self._reloadFilter()
-    }
-
-
-    
     //MARK: UI布局
     private lazy var mContentBGView: UIView = {
         let mContentBGView = UIView()
@@ -216,7 +209,7 @@ class ZXKitLoggerWindow: UIWindow {
                 case 2:
                     self.isSearchViewHidden = false
                 case 3:
-                    self._share()
+                    self.showPicker(pickType: .share)
                 case 4:
                     self.isScrollViewHidden = false
                 case 5:
@@ -233,6 +226,8 @@ class ZXKitLoggerWindow: UIWindow {
                     }
                     let info = "\n" + "current log count".ZXLocaleString + ": \(self.mLogDataArray.count)" +  "\n" + "LogFile count".ZXLocaleString + ": \(count)" + "\n" + "LogFile total size".ZXLocaleString + ": \(size/1024.0)kb"
                     printWarn(info)
+                case 6 :
+                    self.showPicker(pickType: .upload)
                 default:
                     break
             }
@@ -352,6 +347,13 @@ class ZXKitLoggerWindow: UIWindow {
         tView.layer.borderWidth = 1.0
         return tView
     }()
+
+    private lazy var mPickerTipLabel: UILabel = {
+        let tipLabel = UILabel()
+        tipLabel.text = "Please select the log to share".ZXLocaleString
+        tipLabel.font = UIFont.systemFont(ofSize: 14, weight: UIFont.Weight.medium)
+        return tipLabel
+    }()
     
     private lazy var mPickerView: UIPickerView = {
         let tPicker = UIPickerView()
@@ -367,6 +369,45 @@ class ZXKitLoggerWindow: UIWindow {
         tToolBar.barStyle = .default
         return tToolBar
     }()
+}
+
+extension ZXKitLoggerWindow {
+    func insert(model: ZXKitLoggerItem) {
+        if ZXKitLogger.maxDisplayCount != 0 && self.mLogDataArray.count > ZXKitLogger.maxDisplayCount {
+            self.mLogDataArray.removeFirst()
+        }
+        self.mLogDataArray.append(model)
+        if !self.isHidden {
+            self._reloadFilter(model: model)
+        }
+    }
+
+    //
+    func cleanDataArray() {
+        self.mLogDataArray.removeAll()
+        self.mFilterIndexArray.removeAll()
+        self._reloadFilter()
+    }
+
+    private func showPicker(pickType: PickerType) {
+        self.pickerType = pickType
+        self.mFileDateNameList = [String]()
+        let path = HDSqliteTools.shared.getDBFolder()
+        //数据库目录
+        if let enumer = FileManager.default.enumerator(at: path, includingPropertiesForKeys: [URLResourceKey.nameKey], options: [.skipsHiddenFiles, .skipsPackageDescendants]) {
+            while let file = enumer.nextObject() {
+                if let file: URL = file as? URL, file.lastPathComponent.hasSuffix(".db") {
+                    self.mFileDateNameList.append(file.lastPathComponent)
+                }
+            }
+        }
+
+        //倒序，最后的放前面
+        self.mFileDateNameList = self.mFileDateNameList.sorted().reversed()
+        self.mPickerBGView.isHidden = !self.mPickerBGView.isHidden
+        self.mPickerView.reloadAllComponents()
+        self.mShareFileName = self.mFileDateNameList.first ?? ""
+    }
 }
 
 private extension ZXKitLoggerWindow {
@@ -477,47 +518,33 @@ private extension ZXKitLoggerWindow {
 
     @objc private func _confirmPicker() {
         self.mPickerBGView.isHidden = true
-        let dataList = HDSqliteTools.shared.getAllLog(name: self.mShareFileName).reversed()
-        //写入到text文件好解析
-        //文件路径
-        let logFilePathURL = ZXKitUtil.shared.getFileDirectory(type: .caches).appendingPathComponent("ZXKitLogger.log", isDirectory: false)
-        if FileManager.default.fileExists(atPath: logFilePathURL.path) {
-            try? FileManager.default.removeItem(at: logFilePathURL)
-        }
-        do {
-            try dataList.joined(separator: "\n").write(to: logFilePathURL, atomically: false, encoding: String.Encoding.utf8)
-        } catch {
-            print(error)
-        }
-
-        //分享
-        let activityVC = UIActivityViewController(activityItems: [logFilePathURL], applicationActivities: nil)
-        if UIDevice.current.model == "iPad" {
-            activityVC.modalPresentationStyle = UIModalPresentationStyle.popover
-            activityVC.popoverPresentationController?.sourceView = self.mMenuView.subviews.first
-            activityVC.popoverPresentationController?.sourceRect = self.mMenuView.subviews.first?.frame ?? .zero
-        }
-        self._hide()
-        ZXKitUtil.shared.getCurrentVC()?.present(activityVC, animated: true, completion: nil)
-    }
-
-    private func _share() {
-        self.mFileDateNameList = [String]()
-        let path = HDSqliteTools.shared.getDBFolder()
-        //数据库目录
-        if let enumer = FileManager.default.enumerator(at: path, includingPropertiesForKeys: [URLResourceKey.nameKey], options: [.skipsHiddenFiles, .skipsPackageDescendants]) {
-            while let file = enumer.nextObject() {
-                if let file: URL = file as? URL, file.lastPathComponent.hasSuffix(".db") {
-                    self.mFileDateNameList.append(file.lastPathComponent)
-                }
+        if self.pickerType == .share {
+            let dataList = HDSqliteTools.shared.getAllLog(name: self.mShareFileName).reversed()
+            //写入到text文件好解析
+            //文件路径
+            let logFilePathURL = ZXKitUtil.shared.getFileDirectory(type: .caches).appendingPathComponent("ZXKitLogger.log", isDirectory: false)
+            if FileManager.default.fileExists(atPath: logFilePathURL.path) {
+                try? FileManager.default.removeItem(at: logFilePathURL)
             }
-        }
+            do {
+                try dataList.joined(separator: "\n").write(to: logFilePathURL, atomically: false, encoding: String.Encoding.utf8)
+            } catch {
+                print(error)
+            }
 
-        //倒序，最后的放前面
-        self.mFileDateNameList = self.mFileDateNameList.sorted().reversed()
-        self.mPickerBGView.isHidden = !self.mPickerBGView.isHidden
-        self.mPickerView.reloadAllComponents()
-        self.mShareFileName = self.mFileDateNameList.first ?? ""
+            //分享
+            let activityVC = UIActivityViewController(activityItems: [logFilePathURL], applicationActivities: nil)
+            if UIDevice.current.model == "iPad" {
+                activityVC.modalPresentationStyle = UIModalPresentationStyle.popover
+                activityVC.popoverPresentationController?.sourceView = self.mMenuView.subviews.first
+                activityVC.popoverPresentationController?.sourceRect = self.mMenuView.subviews.first?.frame ?? .zero
+            }
+            self._hide()
+            ZXKitUtil.shared.getCurrentVC()?.present(activityVC, animated: true, completion: nil)
+        } else if let complete = ZXKitLogger.uploadComplete {
+            let path = HDSqliteTools.shared.getDBFolder().appendingPathComponent(self.mShareFileName)
+            complete(path)
+        }
     }
 
     //解密
@@ -625,11 +652,8 @@ private extension ZXKitLoggerWindow {
             make.left.right.bottom.equalTo(self.mContentBGView)
         }
 
-        let tipLabel = UILabel()
-        tipLabel.text = "Please select the log to share".ZXLocaleString
-        tipLabel.font = UIFont.systemFont(ofSize: 14, weight: UIFont.Weight.medium)
-        self.mPickerBGView.addSubview(tipLabel)
-        tipLabel.snp.makeConstraints { (make) in
+        self.mPickerBGView.addSubview(mPickerTipLabel)
+        mPickerTipLabel.snp.makeConstraints { (make) in
             make.centerX.equalToSuperview()
             make.top.equalToSuperview().offset(20)
             make.height.equalTo(40)
@@ -637,7 +661,7 @@ private extension ZXKitLoggerWindow {
 
         self.mPickerBGView.addSubview(self.mToolBar)
         self.mToolBar.snp.makeConstraints { (make) in
-            make.top.equalTo(tipLabel.snp.bottom)
+            make.top.equalTo(mPickerTipLabel.snp.bottom)
             make.left.right.equalToSuperview()
             make.height.equalTo(40)
         }
@@ -645,7 +669,7 @@ private extension ZXKitLoggerWindow {
 
         let closeBarItem = UIBarButtonItem(barButtonSystemItem: UIBarButtonItem.SystemItem.cancel, target: self, action: #selector(_closePicker))
         let fixBarItem = UIBarButtonItem(barButtonSystemItem: UIBarButtonItem.SystemItem.flexibleSpace, target: nil, action: nil)
-        let doneBarItem = UIBarButtonItem(title: "Share".ZXLocaleString, style:.plain, target: self, action: #selector(_confirmPicker))
+        let doneBarItem = UIBarButtonItem(title: "Done".ZXLocaleString, style:.plain, target: self, action: #selector(_confirmPicker))
         self.mToolBar.setItems([closeBarItem, fixBarItem, doneBarItem], animated: true)
 
         self.mPickerBGView.addSubview(self.mPickerView)
